+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * Added priority based handling of interrupts in the CPU
+* Improved performance of the interrupt handling
 
 ### Fixed